split cbf codec into many files
