+++ encoding/cbf/cbf_codec/capability.go
@@ -1,0 +1,58 @@
+/*
+ * Cadence - The resource-oriented smart contract programming language
+ *
+ * Copyright 2019-2022 Dapper Labs, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cbf_codec
+
+import (
+	"github.com/onflow/cadence"
+	"github.com/onflow/cadence/encoding/cbf/common_codec"
+)
+
+func (e *Encoder) EncodeCapability(value cadence.Capability) (err error) {
+	err = e.EncodePath(value.Path)
+	if err != nil {
+		return
+	}
+
+	err = common_codec.EncodeAddress(&e.w, value.Address)
+	if err != nil {
+		return
+	}
+
+	return e.EncodeType(value.BorrowType)
+}
+
+func (d *Decoder) DecodeCapability() (capability cadence.Capability, err error) {
+	path, err := d.DecodePath()
+	if err != nil {
+		return
+	}
+
+	address, err := d.DecodeAddress()
+	if err != nil {
+		return
+	}
+
+	borrowType, err := d.DecodeType()
+	if err != nil {
+		return
+	}
+
+	capability = cadence.NewMeteredCapability(d.memoryGauge, path, address, borrowType)
+	return
+}
