+++ encoding/cbf/cbf_codec/character.go
@@ -1,0 +1,40 @@
+/*
+ * Cadence - The resource-oriented smart contract programming language
+ *
+ * Copyright 2019-2022 Dapper Labs, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cbf_codec
+
+import (
+	"github.com/onflow/cadence"
+	"github.com/onflow/cadence/encoding/cbf/common_codec"
+	"github.com/onflow/cadence/runtime/common"
+)
+
+func (d *Decoder) DecodeCharacter() (value cadence.Character, err error) {
+	s, err := common_codec.DecodeString(&d.r)
+	if err != nil {
+		return
+	}
+
+	return cadence.NewMeteredCharacter(
+		d.memoryGauge,
+		common.NewCadenceStringMemoryUsage(len(s)),
+		func() string {
+			return s
+		},
+	)
+}
