+++ encoding/cbf/cbf_codec/dictionary_type.go
@@ -1,0 +1,44 @@
+/*
+ * Cadence - The resource-oriented smart contract programming language
+ *
+ * Copyright 2019-2022 Dapper Labs, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cbf_codec
+
+import (
+	"github.com/onflow/cadence"
+)
+
+func (e *Encoder) EncodeDictionaryType(t cadence.DictionaryType) (err error) {
+	err = e.EncodeType(t.KeyType)
+	if err != nil {
+		return
+	}
+	return e.EncodeType(t.ElementType)
+}
+
+func (d *Decoder) DecodeDictionaryType() (t cadence.DictionaryType, err error) {
+	keyType, err := d.DecodeType()
+	if err != nil {
+		return
+	}
+	elementType, err := d.DecodeType()
+	if err != nil {
+		return
+	}
+	t = cadence.NewMeteredDictionaryType(d.memoryGauge, keyType, elementType)
+	return
+}
