+++ encoding/cbf/cbf_codec/link.go
@@ -1,0 +1,48 @@
+/*
+ * Cadence - The resource-oriented smart contract programming language
+ *
+ * Copyright 2019-2022 Dapper Labs, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cbf_codec
+
+import (
+	"github.com/onflow/cadence"
+	"github.com/onflow/cadence/encoding/cbf/common_codec"
+)
+
+func (e *Encoder) EncodeLink(value cadence.Link) (err error) {
+	err = e.EncodePath(value.TargetPath)
+	if err != nil {
+		return
+	}
+
+	return common_codec.EncodeString(&e.w, value.BorrowType)
+}
+
+func (d *Decoder) DecodeLink() (link cadence.Link, err error) {
+	path, err := d.DecodePath()
+	if err != nil {
+		return
+	}
+
+	borrowType, err := d.DecodeString()
+	if err != nil {
+		return
+	}
+
+	link = cadence.NewMeteredLink(d.memoryGauge, path, string(borrowType))
+	return
+}
