+++ encoding/cbf/cbf_codec/type.go
@@ -36,11 +36,7 @@ func (e *Encoder) EncodeType(t cadence.Type) (err error) {
 		if err != nil {
 			return
 		}
-		err = common_codec.EncodeBool(&e.w, actualType.Type == nil)
-		if err != nil {
-			return
-		}
-		return e.EncodeType(actualType.Type)
+		return e.EncodeOptionalType(actualType)
 	case cadence.StringType:
 		return e.EncodeTypeIdentifier(EncodedTypeString)
 	case cadence.CharacterType:
@@ -110,25 +106,25 @@ func (e *Encoder) EncodeType(t cadence.Type) (err error) {
 		if err != nil {
 			return
 		}
-		return e.EncodeStructType(actualType)
+		return e.EncodeCompositeType(actualType)
 	case *cadence.ResourceType:
 		err = e.EncodeTypeIdentifier(EncodedTypeResource)
 		if err != nil {
 			return
 		}
-		return e.EncodeResourceType(actualType)
+		return e.EncodeCompositeType(actualType)
 	case *cadence.EventType:
 		err = e.EncodeTypeIdentifier(EncodedTypeEvent)
 		if err != nil {
 			return
 		}
-		return e.EncodeEventType(actualType)
+		return e.EncodeCompositeType(actualType)
 	case *cadence.ContractType:
 		err = e.EncodeTypeIdentifier(EncodedTypeContract)
 		if err != nil {
 			return
 		}
-		return e.EncodeContractType(actualType)
+		return e.EncodeCompositeType(actualType)
 	case *cadence.StructInterfaceType:
 		err = e.EncodeTypeIdentifier(EncodedTypeStructInterface)
 		if err != nil {
@@ -231,222 +227,162 @@ func (e *Encoder) EncodeType(t cadence.Type) (err error) {
 	return common_codec.CodecError(fmt.Sprintf("unknown type: %s", t))
 }
 
-func (e *Encoder) EncodeFunctionType(t *cadence.FunctionType) (err error) {
-	err = common_codec.EncodeString(&e.w, t.ID())
+func (d *Decoder) DecodeType() (t cadence.Type, err error) {
+	typeIdentifer, err := d.DecodeTypeIdentifier()
+
 	if err != nil {
 		return
 	}
 
-	err = EncodeArray(e, t.Parameters, e.EncodeParameter)
-	if err != nil {
-		return
+	switch typeIdentifer {
+	case EncodedTypeVoid:
+		t = cadence.NewMeteredVoidType(d.memoryGauge)
+	case EncodedTypeNever:
+		t = cadence.NewMeteredNeverType(d.memoryGauge)
+	case EncodedTypeOptional:
+		t, err = d.DecodeOptionalType()
+	case EncodedTypeBool:
+		t = cadence.NewMeteredBoolType(d.memoryGauge)
+	case EncodedTypeString:
+		t = cadence.NewMeteredStringType(d.memoryGauge)
+	case EncodedTypeCharacter:
+		t = cadence.NewMeteredCharacterType(d.memoryGauge)
+	case EncodedTypeAddress:
+		t = cadence.NewMeteredAddressType(d.memoryGauge)
+	case EncodedTypeNumber:
+		t = cadence.NewMeteredNumberType(d.memoryGauge)
+	case EncodedTypeSignedNumber:
+		t = cadence.NewMeteredSignedNumberType(d.memoryGauge)
+	case EncodedTypeInteger:
+		t = cadence.NewMeteredIntegerType(d.memoryGauge)
+	case EncodedTypeSignedInteger:
+		t = cadence.NewMeteredSignedIntegerType(d.memoryGauge)
+	case EncodedTypeFixedPoint:
+		t = cadence.NewMeteredFixedPointType(d.memoryGauge)
+	case EncodedTypeSignedFixedPoint:
+		t = cadence.NewMeteredSignedFixedPointType(d.memoryGauge)
+	case EncodedTypeInt:
+		t = cadence.NewMeteredIntType(d.memoryGauge)
+	case EncodedTypeInt8:
+		t = cadence.NewMeteredInt8Type(d.memoryGauge)
+	case EncodedTypeInt16:
+		t = cadence.NewMeteredInt16Type(d.memoryGauge)
+	case EncodedTypeInt32:
+		t = cadence.NewMeteredInt32Type(d.memoryGauge)
+	case EncodedTypeInt64:
+		t = cadence.NewMeteredInt64Type(d.memoryGauge)
+	case EncodedTypeInt128:
+		t = cadence.NewMeteredInt128Type(d.memoryGauge)
+	case EncodedTypeInt256:
+		t = cadence.NewMeteredInt256Type(d.memoryGauge)
+	case EncodedTypeUInt128:
+		t = cadence.NewMeteredUInt128Type(d.memoryGauge)
+	case EncodedTypeUInt256:
+		t = cadence.NewMeteredUInt256Type(d.memoryGauge)
+	case EncodedTypeUInt:
+		t = cadence.NewMeteredUIntType(d.memoryGauge)
+	case EncodedTypeUInt8:
+		t = cadence.NewMeteredUInt8Type(d.memoryGauge)
+	case EncodedTypeUInt16:
+		t = cadence.NewMeteredUInt16Type(d.memoryGauge)
+	case EncodedTypeUInt32:
+		t = cadence.NewMeteredUInt32Type(d.memoryGauge)
+	case EncodedTypeUInt64:
+		t = cadence.NewMeteredUInt64Type(d.memoryGauge)
+	case EncodedTypeWord8:
+		t = cadence.NewMeteredWord8Type(d.memoryGauge)
+	case EncodedTypeWord16:
+		t = cadence.NewMeteredWord16Type(d.memoryGauge)
+	case EncodedTypeWord32:
+		t = cadence.NewMeteredWord32Type(d.memoryGauge)
+	case EncodedTypeWord64:
+		t = cadence.NewMeteredWord64Type(d.memoryGauge)
+	case EncodedTypeFix64:
+		t = cadence.NewMeteredFix64Type(d.memoryGauge)
+	case EncodedTypeUFix64:
+		t = cadence.NewMeteredUFix64Type(d.memoryGauge)
+	case EncodedTypeVariableSizedArray:
+		t, err = d.DecodeVariableArrayType()
+	case EncodedTypeConstantSizedArray:
+		t, err = d.DecodeConstantArrayType()
+	case EncodedTypeDictionary:
+		t, err = d.DecodeDictionaryType()
+	case EncodedTypeStruct:
+		t, err = d.DecodeStructType()
+	case EncodedTypeResource:
+		t, err = d.DecodeResourceType()
+	case EncodedTypeEvent:
+		t, err = d.DecodeEventType()
+	case EncodedTypeContract:
+		t, err = d.DecodeContractType()
+	case EncodedTypeStructInterface:
+		t, err = d.DecodeStructInterfaceType()
+	case EncodedTypeResourceInterface:
+		t, err = d.DecodeResourceInterfaceType()
+	case EncodedTypeContractInterface:
+		t, err = d.DecodeContractInterfaceType()
+	case EncodedTypeFunction:
+		t, err = d.DecodeFunctionType()
+	case EncodedTypeReference:
+		t, err = d.DecodeReferenceType()
+	case EncodedTypeRestricted:
+		t, err = d.DecodeRestrictedType()
+	case EncodedTypeBlock:
+		t = cadence.NewMeteredBlockType(d.memoryGauge)
+	case EncodedTypeAuthAccount:
+		t = cadence.NewMeteredAuthAccountType(d.memoryGauge)
+	case EncodedTypePublicAccount:
+		t = cadence.NewMeteredPublicAccountType(d.memoryGauge)
+	case EncodedTypeDeployedContract:
+		t = cadence.NewMeteredDeployedContractType(d.memoryGauge)
+	case EncodedTypeAuthAccountContracts:
+		t = cadence.NewMeteredAuthAccountContractsType(d.memoryGauge)
+	case EncodedTypePublicAccountContracts:
+		t = cadence.NewMeteredPublicAccountContractsType(d.memoryGauge)
+	case EncodedTypeAuthAccountKeys:
+		t = cadence.NewMeteredAuthAccountKeysType(d.memoryGauge)
+	case EncodedTypePublicAccountKeys:
+		t = cadence.NewMeteredPublicAccountKeysType(d.memoryGauge)
+	case EncodedTypeCapabilityPath:
+		t = cadence.NewMeteredCapabilityPathType(d.memoryGauge)
+	case EncodedTypeStoragePath:
+		t = cadence.NewMeteredStoragePathType(d.memoryGauge)
+	case EncodedTypePublicPath:
+		t = cadence.NewMeteredPublicPathType(d.memoryGauge)
+	case EncodedTypePrivatePath:
+		t = cadence.NewMeteredPrivatePathType(d.memoryGauge)
+	case EncodedTypeCapability:
+		var borrowType cadence.Type
+		borrowType, err = d.DecodeType()
+		if err != nil {
+			return
+		}
+		t = cadence.NewMeteredCapabilityType(d.memoryGauge, borrowType)
+	case EncodedTypeEnum:
+		return d.DecodeEnumType()
+
+	case EncodedTypeAnyType:
+		t = cadence.NewMeteredAnyType(d.memoryGauge)
+	case EncodedTypeAnyStructType:
+		t = cadence.NewMeteredAnyStructType(d.memoryGauge)
+	case EncodedTypeAnyResourceType:
+		t = cadence.NewMeteredAnyResourceType(d.memoryGauge)
+	case EncodedTypePath:
+		t = cadence.NewMeteredPathType(d.memoryGauge)
+	case EncodedTypeMetaType:
+		t = cadence.NewMeteredMetaType(d.memoryGauge)
+	default:
+		err = common_codec.CodecError(fmt.Sprintf("unknown type identifier: %d", typeIdentifer))
 	}
-
-	return e.EncodeType(t.ReturnType)
-}
-
-func (e *Encoder) EncodeReferenceType(t cadence.ReferenceType) (err error) {
-	err = common_codec.EncodeBool(&e.w, t.Authorized)
-	if err != nil {
-		return
-	}
-
-	return e.EncodeType(t.Type)
-}
-
-func (e *Encoder) EncodeRestrictedType(t *cadence.RestrictedType) (err error) {
-	err = common_codec.EncodeString(&e.w, t.ID())
-	if err != nil {
-		return
-	}
-
-	err = e.EncodeType(t.Type)
-	if err != nil {
-		return
-	}
-
-	return EncodeArray(e, t.Restrictions, e.EncodeType)
+	return
 }
 
 func (e *Encoder) EncodeTypeIdentifier(id EncodedType) (err error) {
 	return e.writeByte(byte(id))
 }
 
-func (e *Encoder) EncodeVariableArrayType(t cadence.VariableSizedArrayType) (err error) {
-	return e.EncodeType(t.Element())
-}
-
-func (e *Encoder) EncodeConstantArrayType(t cadence.ConstantSizedArrayType) (err error) {
-	err = e.EncodeType(t.Element())
-	if err != nil {
-		return
-	}
-
-	return common_codec.EncodeLength(&e.w, int(t.Size))
-}
-
-func (e *Encoder) EncodeDictionaryType(t cadence.DictionaryType) (err error) {
-	err = e.EncodeType(t.KeyType)
-	if err != nil {
-		return
-	}
-	return e.EncodeType(t.ElementType)
-}
-
-func (e *Encoder) EncodeStructType(t *cadence.StructType) (err error) {
-	err = common_codec.EncodeLocation(&e.w, t.Location)
-	if err != nil {
-		return
-	}
-
-	err = common_codec.EncodeString(&e.w, t.QualifiedIdentifier)
-	if err != nil {
-		return
-	}
-
-	err = EncodeArray(e, t.Fields, func(field cadence.Field) (err error) {
-		return e.EncodeField(field)
-	})
-
-	return EncodeArray(e, t.Initializers, func(parameters []cadence.Parameter) (err error) {
-		return EncodeArray(e, parameters, func(parameter cadence.Parameter) (err error) {
-			return e.EncodeParameter(parameter)
-		})
-	})
-}
-
-func (e *Encoder) EncodeResourceType(t *cadence.ResourceType) (err error) {
-	err = common_codec.EncodeLocation(&e.w, t.Location)
-	if err != nil {
-		return
-	}
-
-	err = common_codec.EncodeString(&e.w, t.QualifiedIdentifier)
-	if err != nil {
-		return
-	}
-
-	err = EncodeArray(e, t.Fields, func(field cadence.Field) (err error) {
-		return e.EncodeField(field)
-	})
-
-	return EncodeArray(e, t.Initializers, func(parameters []cadence.Parameter) (err error) {
-		return EncodeArray(e, parameters, func(parameter cadence.Parameter) (err error) {
-			return e.EncodeParameter(parameter)
-		})
-	})
-}
-
-func (e *Encoder) EncodeEventType(t *cadence.EventType) (err error) {
-	err = common_codec.EncodeLocation(&e.w, t.Location)
-	if err != nil {
-		return
-	}
-
-	err = common_codec.EncodeString(&e.w, t.QualifiedIdentifier)
-	if err != nil {
-		return
-	}
-
-	err = EncodeArray(e, t.Fields, func(field cadence.Field) (err error) {
-		return e.EncodeField(field)
-	})
-
-	return EncodeArray(e, t.Initializer, func(parameter cadence.Parameter) (err error) {
-		return e.EncodeParameter(parameter)
-	})
-}
-
-func (e *Encoder) EncodeContractType(t *cadence.ContractType) (err error) {
-	err = common_codec.EncodeLocation(&e.w, t.Location)
-	if err != nil {
-		return
-	}
-
-	err = common_codec.EncodeString(&e.w, t.QualifiedIdentifier)
-	if err != nil {
-		return
-	}
-
-	err = EncodeArray(e, t.Fields, func(field cadence.Field) (err error) {
-		return e.EncodeField(field)
-	})
-
-	return EncodeArray(e, t.Initializers, func(parameters []cadence.Parameter) (err error) {
-		return EncodeArray(e, parameters, func(parameter cadence.Parameter) (err error) {
-			return e.EncodeParameter(parameter)
-		})
-	})
-}
-
-func (e *Encoder) EncodeInterfaceType(t cadence.InterfaceType) (err error) {
-	err = common_codec.EncodeLocation(&e.w, t.InterfaceTypeLocation())
-	if err != nil {
-		return
-	}
-
-	err = common_codec.EncodeString(&e.w, t.InterfaceTypeQualifiedIdentifier())
-	if err != nil {
-		return
-	}
-
-	err = EncodeArray(e, t.InterfaceFields(), func(field cadence.Field) (err error) {
-		return e.EncodeField(field)
-	})
-
-	return EncodeArray(e, t.InterfaceInitializers(), func(parameters []cadence.Parameter) (err error) {
-		return EncodeArray(e, parameters, func(parameter cadence.Parameter) (err error) {
-			return e.EncodeParameter(parameter)
-		})
-	})
-}
-
-func (e *Encoder) EncodeEnumType(t *cadence.EnumType) (err error) {
-	err = common_codec.EncodeLocation(&e.w, t.Location)
-	if err != nil {
-		return
-	}
-
-	err = common_codec.EncodeString(&e.w, t.QualifiedIdentifier)
-	if err != nil {
-		return
-	}
-
-	err = e.EncodeType(t.RawType)
-	if err != nil {
-		return
-	}
-
-	err = EncodeArray(e, t.Fields, func(field cadence.Field) (err error) {
-		return e.EncodeField(field)
-	})
-
-	return EncodeArray(e, t.Initializers, func(parameters []cadence.Parameter) (err error) {
-		return EncodeArray(e, parameters, func(parameter cadence.Parameter) (err error) {
-			return e.EncodeParameter(parameter)
-		})
-	})
-}
-
-func (e *Encoder) EncodeField(field cadence.Field) (err error) {
-	err = common_codec.EncodeString(&e.w, field.Identifier)
-	if err != nil {
-		return
-	}
-
-	return e.EncodeType(field.Type)
-}
-
-func (e *Encoder) EncodeParameter(parameter cadence.Parameter) (err error) {
-	err = common_codec.EncodeString(&e.w, parameter.Label)
-	if err != nil {
-		return
-	}
-
-	err = common_codec.EncodeString(&e.w, parameter.Identifier)
-	if err != nil {
-		return
-	}
-
-	return e.EncodeType(parameter.Type)
+func (d *Decoder) DecodeTypeIdentifier() (t EncodedType, err error) {
+	b, err := d.read(1)
+	t = EncodedType(b[0])
+	return
 }
